beep_on_error {beepr}	R Documentation

Play a short sound if there is an error

Description

beep_on_error wraps an expression and plays a short sound only if an error occurs.

Usage

beep_on_error(expr, sound = 1)

Arguments

expr
An expression to be evaluated for errors.

sound
character string or number specifying what sound to be played by either specifying one of the built in sounds, specifying the path to a wav file or specifying an url. The default is 1. Possible sounds are:

"ping"

"coin"

"fanfare"

"complete"

"treasure"

"ready"

"shotgun"

"mario"

"wilhelm"

"facebook"

"sword"

If sound does not match any of the sounds above, or is a valid path or url, a random sound will be played. Currently beep can only handle http urls, https is not supported.

Details

If beep is not able to play the sound a warning is issued rather than an error. This is in order to not risk aborting or stopping the process that you wanted to get notified about.

Examples

# Play a "ping" sound if \code{expr} produces an error
beep_on_error(log("foo"))

# Stay silent if \code{expr} does not produce an error
beep_on_error(log(1))

## Not run: 
# Play the Wilhelm scream instead of a ping on error.
beep_on_error(runif("bar"), "wilhelm")

## End(Not run)

[Package beepr version 1.3 Index]