beepr {beepr}	R Documentation

Easily Play Notification Sounds on any Platform

Description

This package contains one function, beep(), with one purpose: To make it easy to play notification sounds on whatever platform you are on. It is intended to be useful, for example, if you are running a long analysis in the background and want to know when it is ready.

Details

The package contains one main function beep, check it out to see what it does. For sound on Windows and MacOS beepr depends on the audio package. For sound on Linux beepr depends on that either the paplay utility from the Pulse Audio system, the aplay utility from the ALSA system, or VLC media player (http://www.videolan.org/vlc/index.html) is installed and on the PATH. Chances are that you already have one of these.

Author(s)

Rasmus Bååth < rasmus.baath@gmail.com >

Examples

# Play a "ping" sound
beep()

[Package beepr version 1.3 Index]